+++ models/staging/stg_fivetran_log__log.sql
@@ -1,10 +1,21 @@
-with log as (
+with base as (
 
     select * 
-    from {{ var('log') }}
+    from {{ ref('stg_fivetran_log__log_tmp') }}
 ),
 
 fields as (
+    select
+        {{
+            fivetran_utils.fill_staging_columns(
+                source_columns=adapter.get_columns_in_relation(ref('stg_fivetran_log__log_tmp')),
+                staging_columns=get_log_columns()
+            )
+        }}
+    from base
+),
+
+final as (
 
     select
         id as log_id, 
@@ -19,8 +30,8 @@ fields as (
         when transformation_id is not null and message_data like '%has failed%' then 'transformation run failed'
         else message_event end as event_subtype,
         transformation_id
-    from log
+    from fields
 )
 
 select * 
-from fields 
+from final