+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Includes the following updates:
 ## Enhancements
 - The `sync_id` field from the source `log` table is added to the `stg_fivetran_log__log` model for ease of grouping events by the sync that they are associated with.
 
+## Under the Hood
+- Added the `get_log_columns` macro and included the fill staging cte's within the `stg_fivetran_log__log` model to ensure the model succeeds regardless of a user not having all the required fields.
+
 ## Documentation Updates
 - The `sync_id` field is added to the documentation in the `fivetran_log.yml` file.